Refactor internal/infra into more targetted packages
